Add a Prefix option in IterationOptions (#628)
When creating an iterator, we pick all the tables in LSM tree. In Dgraph, a lot of times we only need to iterate over all versions of a single key. But, doing a `Seek` costs extra cycles to jump over all the tables which don't have the prefix. This PR adds a `Prefix` option in IteratorOptions, to allow selectively picking only those tables, which could potentially have the `Prefix`.

In benchmarks, I see a 20% improvement in latency with 80 SSTables, when looking for a single key.

Fixes #625 .

Commits:

* Only pick up tables which would have the prefix needed for iteration. This should be faster than picking all of them.
* Left test file behind
* Add benchmark to see the impact of opt.Prefix. I see a 20% improvement when there are 80 SSTables.