Sync files with F_FULLFSYNC on Darwing for Go < 1.12
As documented in https:///golang.org/issues/26650, versions of Go below 1.12
do not use the correct method to sync files on Darwin (MacOS).

This can lead to data loss or corruption, so we should fix it on our codebase.

The proposed solution uses build constraints so this change should only impact
badger on MacOS for versions before Go 1.12.

Fixes #563.