Fix prefix bug in key iterator and allow all versions (#950)
The key iterator is supposed to iterate over all versions of the same
key but the current implementation would iterate over all keys with the 
given prefix. For example, A key iterator for "foo" would return true
for "foo" and "foobar". This commit fixes the issue.

This commit also sets the value of `allVersions` to true in the key 
iterator.

(cherry picked from commit 8b521772ec4ae073883947b26ef6a54b78f459d3)