[breaking/format] Remove vlen from entry header (#945)
This commit removes value length from the entry header stored in each SST.
We don't need vlen to store the length of the value. We can find it by
using the entry offsets stored in the footer of the blocks.

Entries in the table are of the form
+-----------------+-------+------+--------+-----------------+------+
| Klen1 (Point A) | Plen1 | Key1 | Value1 | Klen2 (Point B) | .... |
+-----------------+-------+------+--------+-----------------+------+

And we have the entry index at the end of each block.
+---------------------------+--------------------------+
| Entry1 offset 1 (Point A) | Entry Offset 1 (Point B) |
+---------------------------+--------------------------+

Using the entry index and current position in the buffer, we can find
the length of the value.