Implement  PageBuffer (#996)
PageBuffer can be an alternative for bytes.Buffer. PageBuffer consists of many
pages(fixed size []byte). It starts by writing to the first (empty) page. When it is full,
it allocates next page with double the size of the current page. It does not copy 
content from the current page(as done in bytes.Buffer) and appends it to its 
collection of pages.
It also satisfies writer interface. A reader can be used to read data from it.