Support checksum verification for values read from vlog (#1052)
The values are written with its checksum but we were not validating it
for every read. This commit enables checksum validation for every read.
If "VerifyValueChecksum" is set to true, the checksum will be calculated for
every entry read from the value log. This option has no effect if the
value is stored completely in SST.

The default value of `VerifyValueChecksum` is false.

(cherry picked from commit ecb3b9bb2781414242f6761e0b7e4de74550511f)