Flush vlog buffer if it grows beyond threshold (#1067)
The OOM happens when we trying to flush write batch.
When we insert data into a transaction, we calculate the size
of the transaction by using the following function
https://github.com/dgraph-io/badger/blob/4dc992caecac59dc50481dd2d782d5194119bb0e/txn.go#L295-L300
which uses `estimateSize` function
https://github.com/dgraph-io/badger/blob/4dc992caecac59dc50481dd2d782d5194119bb0e/structs.go#L134-L139

For a key size of 4 and a value size of 1 MB, the `estimateSize`
function would return 4 + 12 + 2 = 18 byte which is valid in case
of SSTs but not for a vlog file. The entry size for vlog would be
around 1 MB and when all such entries are packed into a single
request for vlog, the bytes.buffer in vlog crashes with OOM error.
This commit fixes it.

(cherry picked from commit 9e3dc3f3951a7f676832cf88e705f720f42e5dfa)