Fix level 0 GC dataloss bug (#1090)
When all level 0 tables are kept in memory and GC is triggered,
it is possible that GC removes data from the value log file which
has not yet persisted to disk. This PR fixes this issue by garbage
collecting only the data which has been written to SSTs on the disk.

When L0 is in memory, we GC only the data which is stored in Level 1 onwards.
When L0 is not in memory. we GC data Level 0 onwards.