Fix unsafe usage in Decode (#1097)
We use unsafe pointers to convert slices to structs. It might be
possible that the struct and byte slice are not aligned in the
same way. This PR fixes the alignment issues

The command
```
gotip test -gcflags=all=-d=checkptr
```
no longer complains about unsafe pointer conversion.
The new implementation is as fast as the old one.
Fixes https://github.com/dgraph-io/badger/issues/1096