Fix ValueThreshold for in-memory mode (#1235)
This commit increases the ValueThreshold for in-memory mode. The
existing threshold was of 1 MB which means badger would crash if
len(value) was greater than 1 MB. This commit sets the value threshold
to MaxInt32 (around 2 GB). Badger transaction would return an error if
badger is running in in-memory mode and length of value is greater than
the value threshold.

Fixes https://github.com/dgraph-io/badger/issues/1234