Support multiple iterators in read-write transactions. (#1286)
This adds support for multiple iterators during a read-write transaction. The
iterators created in a read-write transaction will only be able to see writes
that were performed before the iterator was created. Any writes that occur
after the iterator is created will be invisible to the iterator.

Fixes https://github.com/dgraph-io/badger/issues/981