Restore: Account for value size as well (#1358) (#1504)
In the existing code, when performing a restore, we didn't consider the size of the
value. This means badger would consume a lot of memory if the data consisted of
small keys and large values. This PR fixes the issue by forcefully flushing the
buffer if it grows beyond 100 MB.

(cherry picked from commit 02120961faa3cb9706ee9595e4885dfad6679c1f)

Co-authored-by: Ibrahim Jarif <ibrahim@dgraph.io>