perf: Reduce memory usage by better struct packing (#1528)
Badger uses a bunch of short lived structs such as Txn, Item, Entry,
etc. If the members of these structs are not aligned properly, the
compiler will add padding to make them self-aligning. This PR reorders
the struct members to reduce memory footprint and improve performance.

Size of Txn on Master  : 136
Size of Txn on this PR : 120 (11.7% less)

Size of Item on Master  : 168
Size of Item on this PR : 144 (14.2% less)

Size of Entry on Master  : 96
Size of Entry on this PR : 80 (16.7% less)

Size of IteratorOptions on Master  : 56
Size of IteratorOptions on this PR : 40 (28.5% less)

This PR also removed the db from item struct since item.txn already has
access to the db object.
Read http://www.catb.org/esr/structure-packing/ to understand struct packing