opt(memory): Use z.Calloc for allocating KVList  (#1563)
KVs can take up a lot of memory in the stream framework. With this change, we allocate them using z.Allocator, and allow callers to KeyToList to use the allocator to generate KVs as well. After we call Send, we release them.

Also change Stream Framework to spit out StreamDone markers.