fix(compaction): Don't drop data when split overlaps with top tables (#1587)
Badger compactions are split into sub compactions if there are more than 3 tables
bottom tables are involved. The current implementation would drop keys if the top
table would overlap with a bottom table. Consider the following scenario
```
Top table => [C (deleted)]
bot table => [ A ] [ B ] [ C (valid) ] [ D ] 
```
When this range is compacted, the result was `[ A B C(valid) D ] ` 
(only `C(deleted)` was dropped). The expected result was `[A B D]` .

This PR fixes the above-mentioned issue. The test in this PR fails on master
but works with this change.