fix(compaction): Set base level correctly after stream (#1631)
When we use stream writer, all the data is written to the last level. On the restart, 
badger will try to figure out the correct base level. This computation of base level
is incorrect right now because of which we pick the baseLevel which has levels
below it that are empty.