fix(badger): Do not reuse variable across badger commands (#1624) (#1650)
The same variables were being across multiple badger commands. The
default value of a flag that uses the same variable across two commands
could be from one of the commands and it isn't clear which default value
would be used.

The numVersions command was being used in flatten (default value 1)
and stream (default value 0). Running badger stream --help would show
the default value to be 0 but if you print the value of the
numVersions variable, it would turn out to be 1. This is misleading.

This PR separates the variable so that they don't overlap.

This PR also adds the compression flag to the flatten tool.

(cherry picked from commit bd5a05befc0f2df077a7f3b47546de8006f19812)

Co-authored-by: Ibrahim Jarif <ibrahim@dgraph.io>