fix(compaction): Set base level correctly after stream (#1631) (#1651)
When we use stream writer, all the data is written to the last level. On the restart,
badger will try to figure out the correct base level. This computation of base level
is incorrect right now because of which we pick the baseLevel which has levels
below it that are empty.

(cherry picked from commit f9c1eeb36a0ab4fe961b25e5c70a189374c302ca)

Co-authored-by: Ibrahim Jarif <ibrahim@dgraph.io>