fix(pendingWrites): don't skip the pending entries with version=0 (#1721)
With the introduction of SinceTs, a bug was introduced #1653 that skips the pending entries.
The default value of SinceTs is zero. And for the transaction made at readTs 0, the pending entries have version set to 0. So they were also getting skipped.

(cherry picked from commit 50276c44e257f3a655f4747cce6116359aeecb7c)