feat(zstd): backport replacement of DataDog's zstd with Klauspost's zstd (#1736)
Backports the changes from #1706 and #1709 to the v2 branch of badger.

For context, https://github.com/caddyserver/caddy uses https://github.com/smallstep/certificates which uses https://github.com/smallstep/nosql which uses badger.

See smallstep/nosql#12 for the discussion. It's not currently viable for smallstep to upgrade from v2 to v3 of badger, so it would be nice to backport this zstd change to the v2 branch to remove the implicit (yet optional) CGO dependency on Caddy.

I targeted the release/v2.2007 branch, I couldn't really see a more relevant branch to target, so I hope that's okay.