fix(test): The new test case PagebufferReader5 introduced an error. (#1936)
## Problem
Fixes DGRAPHCORE-199
The intended fix for the CI issue for randomly getting EOF failures in
PagebufferReader2 test introduced the test case FixPagebufferReader5 --
intended to avoid regressions. However, it did something that is not
expected by the PageBuffer struct: it created a new PageBufferReader
based on an empty PageBuffer. It seems that this is not allowed.
 
## Solution
Fix the test case for PagebufferReader5 by no longer attempting to
create a new PageBufferReader based on an empty PageBuffer.

### Note
The CI build will still fail to allow this to pass because it will
continue to use the prior version of the tests without the fix. However,
locally tested, I get this result:
~/gitspace/github.com/dgraph-io/badger/y$ go test -run PagebufferReader5
PASS
ok  	github.com/dgraph-io/badger/v4/y	0.335s

---------

Co-authored-by: skrdgraph <sudhish@dgraph.io>
Co-authored-by: Joshua Goldstein <joshua@dgraph.io>