fix crash loop on missing manifest tables (#2198)
Please see
https://discuss.hypermode.com/t/manifest-removes-non-existing-table/19882

We are experiencing an error where dgraph tries to remove a table that is
missing. I assume that given it has already been removed, it should not
trigger a crash loop in compaction. Instead i sense check that the table
is indeed removed from all levels, and return without an error code.